Updated the Book.java, change Long to int

+++ src/main/java/com/kaluzny/domain/BookRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import java.util.List;
 
 @RepositoryRestResource
-public interface BookRepository extends MongoRepository<Book, Long> {
+public interface BookRepository extends MongoRepository<Book, Integer> {
 
     List<Book> findByName(String name);
 }
